+++ src/ripple/test/JTx.h
@@ -1,26 +1,30 @@
 //------------------------------------------------------------------------------
 /*
-  This file is part of rippled: https://github.com/ripple/rippled
-  Copyright (c) 2012-2015 Ripple Labs Inc.
+    This file is part of rippled: https://github.com/ripple/rippled
+    Copyright (c) 2012, 2013 Ripple Labs Inc.
 
-  Permission to use, copy, modify, and/or distribute this software for any
-  purpose  with  or without fee is hereby granted, provided that the above
-  copyright notice and this permission notice appear in all copies.
+    Permission to use, copy, modify, and/or distribute this software for any
+    purpose  with  or without fee is hereby granted, provided that the above
+    copyright notice and this permission notice appear in all copies.
 
-  THE  SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
-  WITH  REGARD  TO  THIS  SOFTWARE  INCLUDING  ALL  IMPLIED  WARRANTIES  OF
-  MERCHANTABILITY  AND  FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
-  ANY  SPECIAL ,  DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
-  WHATSOEVER  RESULTING  FROM  LOSS  OF USE, DATA OR PROFITS, WHETHER IN AN
-  ACTION  OF  CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
-  OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+    THE  SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+    WITH  REGARD  TO  THIS  SOFTWARE  INCLUDING  ALL  IMPLIED  WARRANTIES  OF
+    MERCHANTABILITY  AND  FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+    ANY  SPECIAL ,  DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+    WHATSOEVER  RESULTING  FROM  LOSS  OF USE, DATA OR PROFITS, WHETHER IN AN
+    ACTION  OF  CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+    OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 */
 //==============================================================================
 
 #ifndef RIPPLE_APP_TESTS_JTX_H_INCLUDED
 #define RIPPLE_APP_TESTS_JTX_H_INCLUDED
 
-#include <ripple/app/tests/Common.h>
+#include <ripple/test/Account.h>
+#include <ripple/test/amounts.h>
+#include <ripple/test/jtx/any.h>
+#include <ripple/test/jtx/json.h>
+#include <ripple/test/jtx/tags.h>
 #include <ripple/app/ledger/Ledger.h>
 #include <ripple/json/json_value.h>
 #include <ripple/protocol/Issue.h>
@@ -89,80 +93,6 @@ struct JTx
 
 namespace jtx {
 
-//
-// Dispatch Tags
-//
-
-struct none_t { none_t() { } };
-static none_t const none;
-
-struct autofill_t { autofill_t() { } };
-static autofill_t const autofill;
-
-struct disabled_t { disabled_t() { } };
-static disabled_t const disabled;
-
-//
-// Helpers
-//
-
-struct MaybeAnyAmount;
-
-struct any_t
-{
-    any_t() { }
-
-    inline
-    MaybeAnyAmount
-    operator()(STAmount const& sta) const;
-};
-
-// This wrapper helps pay destinations
-// in their own issue using generic syntax
-struct MaybeAnyAmount
-{
-    bool is_any;
-    STAmount value;
-
-    MaybeAnyAmount() = delete;
-    MaybeAnyAmount (MaybeAnyAmount const&) = default;
-    MaybeAnyAmount& operator= (MaybeAnyAmount const&) = default;
-
-    MaybeAnyAmount (STAmount const& amount)
-        : is_any(false)
-        , value(amount)
-    {
-    }
-
-    MaybeAnyAmount (STAmount const& amount,
-            any_t const*)
-        : is_any(true)
-        , value(amount)
-    {
-    }
-
-    // Reset the issue to a specific account
-    void
-    to (ripple::Account const& id)
-    {
-        if (! is_any)
-            return;
-        value.setIssuer(id);
-    }
-};
-
-inline
-MaybeAnyAmount
-any_t::operator()(STAmount const& sta) const
-{
-    return MaybeAnyAmount(sta, this);
-}
-
-/** Returns an amount representing "any issuer"
-    @note With respect to what the recipient will accept
-*/
-static any_t const any;
-
 //------------------------------------------------------------------------------
 //
 // Utilities
@@ -204,64 +134,6 @@ struct parse_error : std::logic_error
 */
 STObject
 parse (Json::Value const& jv);
-
-//------------------------------------------------------------------------------
-//
-// JSON generators
-//
-//------------------------------------------------------------------------------
-
-/** Add and/or remove flag. */
-Json::Value
-fset (Account const& account,
-    std::uint32_t on, std::uint32_t off = 0);
-
-/** Remove account flag. */
-inline
-Json::Value
-fclear (Account const& account,
-    std::uint32_t off)
-{
-    return fset(account, 0, off);
-}
-
-/** Create a payment. */
-Json::Value
-pay (Account const& account,
-    Account const& to,
-        MaybeAnyAmount amount);
-
-/** Create an offer. */
-Json::Value
-offer (Account const& account,
-    STAmount const& in, STAmount const& out);
-
-/** Set a transfer rate. */
-Json::Value
-rate (Account const& account,
-    double multiplier);
-
-/** Disable the regular key. */
-Json::Value
-regkey (Account const& account,
-    disabled_t);
-
-/** Set a regular key. */
-Json::Value
-regkey (Account const& account,
-    Account const& signer);
-/** The null transaction. */
-inline
-Json::Value
-noop (Account const& account)
-{
-    return fset(account, 0);
-}
-
-/** Modify a trust line. */
-Json::Value
-trust (Account const& account,
-    STAmount const& amount);
 
 //------------------------------------------------------------------------------
 //
