+++ src/ripple/test/impl/amounts.cpp
@@ -18,7 +18,7 @@
 //==============================================================================
 
 #include <BeastConfig.h>
-#include <ripple/app/tests/Common.h>
+#include <ripple/test/amounts.h>
 #include <ripple/protocol/SystemParameters.h>
 
 namespace ripple {
@@ -57,62 +57,6 @@ STAmount
 IOU::operator()(detail::epsilon_multiple m) const
 {
     return STAmount(issue_, m.n, -81);
-}
-
-//------------------------------------------------------------------------------
-
-#ifdef _MSC_VER
-Account::Account (Account&& other)
-    : name_(std::move(other.name_))
-    , pk_(std::move(other.pk_))
-    , sk_(std::move(other.sk_))
-    , id_(std::move(other.id_))
-    , human_(std::move(other.human_))
-{
-}
-
-Account&
-Account::operator= (Account&& rhs)
-{
-    name_ = std::move(rhs.name_);
-    pk_ = std::move(rhs.pk_);
-    sk_ = std::move(rhs.sk_);
-    id_ = std::move(rhs.id_);
-    human_ = std::move(rhs.human_);
-    return *this;
-}
-#endif
-
-Account::Account(
-        std::string name, KeyPair&& keys)
-    : name_(std::move(name))
-{
-    pk_ = std::move(keys.publicKey);
-    sk_ = std::move(keys.secretKey);
-    id_ = pk_.getAccountID();
-    human_ = pk_.humanAccountID();
-}
-
-Account::Account (std::string name,
-        KeyType type)
-#ifndef _MSC_VER
-    : Account(name,
-#else
-    // Fails on Clang and possibly gcc
-    : Account(std::move(name),
-#endif
-        generateKeysFromSeed(type,
-            RippleAddress::createSeedGeneric(
-                name)))
-{
-}
-
-IOU
-Account::operator[](std::string const& s) const
-{
-    auto const currency = to_currency(s);
-    assert(currency != noCurrency());
-    return IOU(Issue(currency, id()));
 }
 
 } // test
